pyproject.toml
[project]
name = "custom-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
]
